move after and plugin_config into init.lua
