--- nvim/.config/nvim/after/ftplugin/help.lua
@@ -1,2 +1,0 @@
-vim.opt_local.colorcolumn = ""
-vim.opt_local.cursorcolumn = false
