--- nvim/.config/nvim/after/ftplugin/javascript.lua
@@ -1,2 +1,0 @@
-vim.cmd("iab <buffer> tbitd toBeInTheDocument()")
-vim.cmd("iab <buffer> fna () => {}")
