--- nvim/.config/nvim/after/ftplugin/javascriptreact.lua
@@ -1,1 +1,0 @@
-vim.cmd("runtime! ftplugin/javascript.lua")
