--- nvim/.config/nvim/after/ftplugin/typescript.lua
@@ -1,1 +1,0 @@
-vim.cmd("runtime! ftplugin/javascript.lua")
