--- nvim/.config/nvim/after/ftplugin/typescriptreact.lua
@@ -1,1 +1,0 @@
-vim.cmd("runtime! ftplugin/javascriptreact.lua")
