--- nvim/.config/nvim/lua/plugin_config/other.lua
@@ -1,43 +1,0 @@
--- vim-tmux-navigator, no setup call required
-vim.g.tmux_navigator_no_wrap = 1
-vim.keymap.set("n", "<c-h>", "<cmd>TmuxNavigateLeft<cr>")
-vim.keymap.set("n", "<c-j>", "<cmd>TmuxNavigateDown<cr>")
-vim.keymap.set("n", "<c-k>", "<cmd>TmuxNavigateUp<cr>")
-vim.keymap.set("n", "<c-l>", "<cmd>TmuxNavigateRight<cr>")
-
--- nvim-surround, no config required
-require("nvim-surround").setup()
-
--- conform
-require("conform").setup({
-	formatters_by_ft = {
-		javascript = { "prettierd" },
-		typescript = { "prettierd" },
-		javascriptreact = { "prettierd" },
-		typescriptreact = { "prettierd" },
-		css = { "prettierd" },
-		html = { "prettierd" },
-		json = { "prettierd" },
-		yaml = { "prettierd" },
-		markdown = { "prettierd" },
-		lua = { "stylua" },
-	},
-	format_on_save = {
-		lsp_fallback = true,
-		timeout_ms = 500,
-	},
-})
-
--- treesitter
-local parsers = { "comment", "css", "javascript", "lua", "typescript", "tsx", "vim", "vimdoc" }
-
----@diagnostic disable-next-line: missing-fields
-require("nvim-treesitter.configs").setup({
-	ensure_installed = parsers,
-	sync_install = false,
-	auto_install = false,
-	highlight = {
-		enable = true,
-		additional_vim_regex_highlighting = false,
-	},
-})
